refactor: 사용하지 않는 클래스제거. HttpRequestMessageReader 를 private class 로 변경
