--- src/main/java/webserver/HttpHeaderParser.java
@@ -1,4 +1,0 @@
-package webserver;
-
-public interface HttpHeaderParser {
-}
